python-to-node/python-grpc-client/pyproject.toml
[project]
name = "python-grpc-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpcio>=1.80.0",
    "grpcio-tools>=1.80.0",
]
